Backend/PersonalBlog/requirements.txt
Django
psycopg2-binary
dj-rest-auth
django-allauth
django-cors-headers
django-filter
djangorestframework
djangorestframework-simplejwt
pillow
channels
daphne
gunicorn
python-dotenv
drf-spectacular
PyJWT